create ServiceBaseTestCase for service class .
