+++ app/backend/tests/Unit/Service/RolesServiceTest.php
@@ -4,6 +4,7 @@ namespace Tests\Unit\Service;
 
 // use PHPUnit\Framework\TestCase;
 use Tests\TestCase;
+use Tests\ServiceBaseTestCase;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Foundation\Testing\WithFaker;
@@ -16,45 +17,8 @@ use Database\Seeders\PermissionsTableSeeder;
 use Database\Seeders\RolePermissionsTableSeeder;
 use Database\Seeders\RolesTableSeeder;
 
-class RolesServiceTest extends TestCase
+class RolesServiceTest extends ServiceBaseTestCase
 {
-    use HelperTrait;
-
-    protected $initialized = false;
-
-    /**
-     * 初期化処理
-     *
-     * @return array
-     */
-    protected function init(): array
-    {
-
-        // 'testing'
-        // $env = Config::get('app.env');
-
-        Artisan::call('migrate:fresh');
-        // Artisan::call('db:seed');
-        $this->seed([
-            AdminsTableSeeder::class,
-            PermissionsTableSeeder::class,
-            RolesTableSeeder::class,
-            RolePermissionsTableSeeder::class,
-            AdminsRolesTableSeeder::class,
-        ]);
-
-        $response = $this->json('POST', route('auth.admin.login'), [
-            'email'    => Config::get('myappTest.test.admin.login.email'),
-            'password' => Config::get('myappTest.test.admin.login.password')
-        ])->json();
-
-        return [
-            'token'          => $response['access_token'],
-            'user_id'        => $response['user']['id'],
-            'user_authority' => $response['user']['authority']
-        ];
-    }
-
     /**
      * setUpは各テストメソッドが実行される前に実行する
      * 親クラスのsetUpを必ず実行する
